backend/requirements.txt
fastapi
uvicorn
pytest
httpx
python-multipart
python-dotenv
psycopg2
